ui: live progress bar, centered layout, real-model verified
- Stream real progress into the Status box + a proper gradient progress bar
  (indeterminate shimmer while the model loads, determinate "N / total · %"
  during translation) via a generator that drains a worker-thread queue.
  Replaces Gradio's bare spinner/elapsed-seconds indicator.
- Center the app container (fix left-aligned layout on wide windows).
- Drop the external Google Fonts @import (it blocked first paint and kept the
  boot loader up); headings now use a system font stack.
- translator: transformers 5 renamed torch_dtype -> dtype.
- Surface model/runtime errors from the worker thread as a gr.Error.

Verified end-to-end against the real fine-tuned model: correct Sinhala output,
credit lines skipped, streamed progress, and a downloadable .srt.

+++ srt-app/translator.py
@@ -145,10 +145,11 @@ def _load_model_cached(model_dir: str):
 
     tokenizer = AutoTokenizer.from_pretrained(str(model_path), local_files_only=True)
     dtype = torch.float16 if torch.cuda.is_available() else torch.float32
+    # transformers 5 renamed torch_dtype -> dtype.
     model = AutoModelForSeq2SeqLM.from_pretrained(
         str(model_path),
         local_files_only=True,
-        torch_dtype=dtype,
+        dtype=dtype,
     )
     device = "cuda" if torch.cuda.is_available() else "cpu"
     model.to(device)
